Create repository and user fake db generator, and create level class
